Refactor: ProfessorCourseService에 CQRS 패턴 적용 #1
